Calculate the conditional Mahalanobis distance for any variables.

Calculate the conditional Mahalanobis distance for any variables.

CMahalanobis(Dep, Ind = NULL, R, d)

Arguments

Dep
The names of variables you would like to condition on.

Ind
The names of variables of your interest.

R
Conditional correlation among variables.

d
Profiles of interest.

Value

conditional Mahalanobis distance, percentiles for each case based on the Chi-square distribution formed by conditional Mahalanobis distance and predicted Deps based on Inds.

Examples

SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
d_demo <- simStandardized(SimModel, 10)
CMahalanobis(c("Math", "Read"),c("Gf", "Gs", "Gc"),d_demo$R_all,d_demo$Data)

#> $dCM
#>             [,1]
#>  [1,] 0.23504126
#>  [2,] 0.44103057
#>  [3,] 0.20302835
#>  [4,] 2.38264613
#>  [5,] 2.51773899
#>  [6,] 3.87974379
#>  [7,] 0.09262599
#>  [8,] 0.51730923
#>  [9,] 0.25726965
#> [10,] 1.58137201
#> 
#> $df
#> [1] 2
#> 
#> $p
#>             [,1]
#>  [1,] 0.11087783
#>  [2,] 0.19789462
#>  [3,] 0.09653163
#>  [4,] 0.69618097
#>  [5,] 0.71602512
#>  [6,] 0.85627764
#>  [7,] 0.04525692
#>  [8,] 0.22791036
#>  [9,] 0.12070500
#> [10,] 0.54646644
#> 
#> $PredictedSubtests
#>              Math        Read
#>  [1,]  0.49043328  0.32682991
#>  [2,] -0.18084607 -0.14827959
#>  [3,] -0.06598051 -0.24144725
#>  [4,] -0.62418001 -0.68377288
#>  [5,]  1.05642756  0.80333276
#>  [6,] -0.33224812 -0.17388008
#>  [7,]  0.43920928  0.52134446
#>  [8,] -0.28817079 -0.26466975
#>  [9,] -0.26588496 -0.08454291
#> [10,] -0.17697527  0.12212633
#> 
#> $R2
#>  Math  Read 
#> 0.264 0.222 
#>