Calculate the conditional Mahalanobis distance based on factor scores.

Calculate the conditional Mahalanobis distance based on factor scores.

CMahalanobis_FS(Dep, Ind = NULL, R, d)

Arguments

Dep
The names of variables you would like to condition on.

Ind
The names of variables of your interest.

R
Conditional correlation among variables.

d
Profiles of interest.

Value

conditional Mahalanobis distance, percentiles for each case based on the Chi-square distribution formed by conditional Mahalanobis distance and predicted Deps based on Inds.

Examples

SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
d_demo <- simStandardized(SimModel, 10)
CMahalanobis_FS(c("Math", "Read"),c("Gf", "Gs", "Gc"),d_demo$R_all,d_demo$Data)

#> $dCM
#>            [,1]
#>  [1,] 1.5124978
#>  [2,] 0.4034243
#>  [3,] 0.2778164
#>  [4,] 1.0282637
#>  [5,] 2.1021020
#>  [6,] 1.3500481
#>  [7,] 0.7102342
#>  [8,] 2.4659680
#>  [9,] 0.4614183
#> [10,] 3.0250635
#> 
#> $df
#> [1] 2
#> 
#> $p
#>            [,1]
#>  [1,] 0.5305760
#>  [2,] 0.1826698
#>  [3,] 0.1296921
#>  [4,] 0.4019805
#>  [5,] 0.6504298
#>  [6,] 0.4908558
#>  [7,] 0.2989087
#>  [8,] 0.7085783
#>  [9,] 0.2060296
#> [10,] 0.7796486
#> 
#> $PredictedSubtests
#>              Math        Read
#>  [1,] -0.03639824  0.28167494
#>  [2,] -0.23387201 -0.11586284
#>  [3,] -0.31717382 -0.35566309
#>  [4,]  0.44356618  0.62837074
#>  [5,] -0.02302250  0.06227216
#>  [6,]  0.51074623  0.73308791
#>  [7,] -0.70455607 -0.56499788
#>  [8,]  0.30585453  0.31161875
#>  [9,] -0.37871102 -0.09379568
#> [10,]  0.60050925  0.25396571
#> 
#> $R2
#>  Math  Read 
#> 0.264 0.222 
#>