Function to evaluate the accuracy of dCM with estimated factor scores

Function to evaluate the accuracy of dCM with estimated factor scores

dcm_cor(model, v_dep, v_ind, n = 10000)

Arguments

model
Lavaan Syntax Object

v_dep
The names of variables you would like to condition on.

v_ind
The names of variables of your interest.

n
Sample size of simulated data

Value

Correlation between the conditional Mahalanobis distance calculated by using the true scores and the conditional Mahalanobis calculated by using estimated factor scores

Examples

m <- "
Gc =~ 0.85 * Gc_1 + 0.68 * Gc_2 + 0.80 * Gc_3
Gf =~ 0.80 * Gf_1 + 0.90 * Gf_2 + 0.80 * Gf_3
Gs =~ 0.70 * Gs_1 + 0.80 * Gs_2 + 0.80 * Gs_3
Read =~ 0.66 * Read_1 + 0.85 * Read_2 + 0.91 * Read_3
Math =~ 0.40 * Math_1 + 0.90 * Math_2 + 0.70 * Math_3
Gc ~ 0.60 * Gf + 0.10 * Gs
Gf ~ 0.50 * Gs
Read ~ 0.40 * Gc + 0.10 * Gf
Math ~ 0.20 * Gc + 0.30 * Gf + 0.10 * Gs
"
dcm_cor(m, v_dep = c("Math", "Read"), v_ind = c("Gc", "Gf", "Gs"), n = 100)

#> [1] 0.6912473