Estimate factor scores for a given profile and population model.

Estimate factor scores for a given profile and population model.

estStandardized(d, m)

Arguments

d
observed z-scores in matrix or data.frame

m
Structural model represented by lavaan Syntax

Value

simStandardized output list with estimated factor scores from d

Examples

m = "latent_1 =~ 0.8 * ob_1 + 0.8 * ob_2 + 0.8 * ob_3"
d <- data.frame(ob_1 = 1, ob_2 = -0.5, ob_3 = 1.2)

# Estimate factor scores based on this case
estStandardized(d = d, m = m)

#> $Data
#>   ob_1 ob_2 ob_3 latent_1_FS e_ob_1_FS e_ob_2_FS e_ob_3_FS
#> 1    1 -0.5  1.2   0.5964912  0.871345 -1.628655  1.204678
#> 
#> $vObserved
#> [1] "ob_1" "ob_2" "ob_3"
#> 
#> $vError
#> [1] "e_ob_1" "e_ob_2" "e_ob_3"
#> 
#> $R
#>          latent_1 ob_1 ob_2 ob_3 e_ob_1 e_ob_2 e_ob_3
#> latent_1      1.0 0.80 0.80 0.80    0.0    0.0    0.0
#> ob_1          0.8 1.00 0.64 0.64    0.6    0.0    0.0
#> ob_2          0.8 0.64 1.00 0.64    0.0    0.6    0.0
#> ob_3          0.8 0.64 0.64 1.00    0.0    0.0    0.6
#> e_ob_1        0.0 0.60 0.00 0.00    1.0    0.0    0.0
#> e_ob_2        0.0 0.00 0.60 0.00    0.0    1.0    0.0
#> e_ob_3        0.0 0.00 0.00 0.60    0.0    0.0    1.0
#> 
#> $A
#>          latent_1 ob_1 ob_2 ob_3
#> latent_1      0.0    0    0    0
#> ob_1          0.8    0    0    0
#> ob_2          0.8    0    0    0
#> ob_3          0.8    0    0    0
#> 
#> $S
#>          latent_1 ob_1 ob_2 ob_3
#> latent_1        1 0.00 0.00 0.00
#> ob_1            0 0.36 0.00 0.00
#> ob_2            0 0.00 0.36 0.00
#> ob_3            0 0.00 0.00 0.36
#> 
#> $iterations
#> [1] 2
#> 
#> $vLatent
#> [1] "latent_1"
#> 
#> $vDisturbance
#> character(0)
#> 
#> $vFactorScores
#> [1] "latent_1_FS" "e_ob_1_FS"   "e_ob_2_FS"   "e_ob_3_FS"  
#> 
#> $FactorScoreCoef
#>       latent_1     e_ob_1     e_ob_2     e_ob_3
#> ob_1 0.3508772  1.1988304 -0.4678363 -0.4678363
#> ob_2 0.3508772 -0.4678363  1.1988304 -0.4678363
#> ob_3 0.3508772 -0.4678363 -0.4678363  1.1988304
#> 
#> $R_all
#>              latent_1      ob_1         ob_2      ob_3     e_ob_1     e_ob_2
#> latent_1    1.0000000 0.8000000 8.000000e-01 0.8000000  0.0000000  0.0000000
#> ob_1        0.8000000 1.0000000 6.400000e-01 0.6400000  0.6000000  0.0000000
#> ob_2        0.8000000 0.6400000 1.000000e+00 0.6400000  0.0000000  0.6000000
#> ob_3        0.8000000 0.6400000 6.400000e-01 1.0000000  0.0000000  0.0000000
#> e_ob_1      0.0000000 0.6000000 0.000000e+00 0.0000000  1.0000000  0.0000000
#> e_ob_2      0.0000000 0.0000000 6.000000e-01 0.0000000  0.0000000  1.0000000
#> e_ob_3      0.0000000 0.0000000 0.000000e+00 0.6000000  0.0000000  0.0000000
#> latent_1_FS 0.9176629 0.8717798 8.717798e-01 0.8717798  0.2294157  0.2294157
#> e_ob_1_FS   0.2482286 0.7074516 6.545242e-17 0.0000000  0.8481145 -0.3309715
#> e_ob_2_FS   0.2482286 0.0000000 7.074516e-01 0.0000000 -0.3309715  0.8481145
#> e_ob_3_FS   0.2482286 0.0000000 0.000000e+00 0.7074516 -0.3309715 -0.3309715
#>                 e_ob_3 latent_1_FS     e_ob_1_FS  e_ob_2_FS  e_ob_3_FS
#> latent_1     0.0000000   0.9176629  2.482286e-01  0.2482286  0.2482286
#> ob_1         0.0000000   0.8717798  7.074516e-01  0.0000000  0.0000000
#> ob_2         0.0000000   0.8717798  6.545242e-17  0.7074516  0.0000000
#> ob_3         0.6000000   0.8717798  0.000000e+00  0.0000000  0.7074516
#> e_ob_1       0.0000000   0.2294157  8.481145e-01 -0.3309715 -0.3309715
#> e_ob_2       0.0000000   0.2294157 -3.309715e-01  0.8481145 -0.3309715
#> e_ob_3       1.0000000   0.2294157 -3.309715e-01 -0.3309715  0.8481145
#> latent_1_FS  0.2294157   1.0000000  2.705009e-01  0.2705009  0.2705009
#> e_ob_1_FS   -0.3309715   0.2705009  1.000000e+00 -0.3902439 -0.3902439
#> e_ob_2_FS   -0.3309715   0.2705009 -3.902439e-01  1.0000000 -0.3902439
#> e_ob_3_FS    0.8481145   0.2705009 -3.902439e-01 -0.3902439  1.0000000
#> 
#> $R_FS
#>           latent_1     e_ob_1     e_ob_2     e_ob_3
#> latent_1 1.0000000  0.2705009  0.2705009  0.2705009
#> e_ob_1   0.2705009  1.0000000 -0.3902439 -0.3902439
#> e_ob_2   0.2705009 -0.3902439  1.0000000 -0.3902439
#> e_ob_3   0.2705009 -0.3902439 -0.3902439  1.0000000
#> 
#> $FactorScoreValidity
#>  latent_1    e_ob_1    e_ob_2    e_ob_3 
#> 0.8421053 0.7192982 0.7192982 0.7192982 
#> 
#> $FactorScoreSE
#> se.latent_1   se.e_ob_1   se.e_ob_2   se.e_ob_3 
#>   0.3973597   0.5298129   0.5298129   0.5298129 
#> 
#> $Model
#> [1] "latent_1 =~ 0.8 * ob_1 + 0.8 * ob_2 + 0.8 * ob_3"
#>