Wrapper for finding out mahalanobis distance between variables: this one gives everything for practioners to use when they only have population relatons and their clients' data

Wrapper for finding out mahalanobis distance between variables: this one gives everything for practioners to use when they only have population relatons and their clients' data

maha(Model, Dep, Ind = NULL, d)

Arguments

Model
Population relations among variables represented by Lavaan Syntax.

Dep
The names of variables you would like to condition on.

Ind
The names of variables of your interest.

d
Profiles of interest.

Value

conditional Mahalanobis distance, percentiles for each case based on the Chi-square distribution formed by conditional Mahalanobis distance and predicted Deps based on Inds.

Examples

SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
d_demo <- simStandardized(SimModel, 10)
maha(SimModel, Dep = c("Math", "Read"),Ind = c("Gc", "Gf", "Gs") , d_demo$Data)

#> $dCM
#>            [,1]
#>  [1,] 3.4626504
#>  [2,] 1.7070532
#>  [3,] 0.2336019
#>  [4,] 2.7535081
#>  [5,] 0.2790888
#>  [6,] 0.1625096
#>  [7,] 1.3642667
#>  [8,] 4.8319535
#>  [9,] 6.3683101
#> [10,] 2.4599385
#> 
#> $df
#> [1] 2
#> 
#> $p
#>             [,1]
#>  [1,] 0.82295037
#>  [2,] 0.57408973
#>  [3,] 0.11023770
#>  [4,] 0.74760351
#>  [5,] 0.13024558
#>  [6,] 0.07804125
#>  [7,] 0.49446264
#>  [8,] 0.91071991
#>  [9,] 0.95858678
#> [10,] 0.70769843
#> 
#> $PredictedSubtests
#>              Math        Read
#>  [1,] -0.24869608 -0.30524959
#>  [2,] -0.39121051 -0.44881867
#>  [3,] -0.43806524 -0.55983428
#>  [4,] -0.08537076 -0.14295976
#>  [5,] -0.02715368 -0.27329733
#>  [6,]  0.25517797  0.31771699
#>  [7,] -0.22484976  0.04011313
#>  [8,]  0.23338601  0.25883246
#>  [9,]  0.41186324  0.06505047
#> [10,] -0.21875697 -0.09826840
#> 
#> $R2
#>  Math  Read 
#> 0.264 0.222 
#>