function to make a set of parameters that follows beta distribution

function to make a set of parameters that follows beta distribution

makePara(from, to, n, k)

Arguments

from
starting point for the parameters

to
ending point for the parameters

n
number of cases

k
number of variables.

Value

all parameters, lavaan object with specified parameters and data genereged for each condition by calling simStandarized and the accuracy index for each ondition

Examples

makePara(1000, 9000, n = 10, k = 15)

#>            X1        X2        X3         X4        X5        X6        X7
#> 1  0.09742837 0.1042426 0.1040001 0.09517797 0.1016401 0.1013950 0.1065240
#> 2  0.18586168 0.1868326 0.1913607 0.18952038 0.1933241 0.1855822 0.1853229
#> 3  0.28112051 0.2822426 0.2742311 0.28836416 0.2838639 0.2739212 0.2810932
#> 4  0.36715692 0.3690494 0.3747573 0.37331306 0.3622045 0.3649523 0.3710217
#> 5  0.45523907 0.4591878 0.4554997 0.45732703 0.4577638 0.4530817 0.4528795
#> 6  0.55349102 0.5481690 0.5469658 0.54690558 0.5554864 0.5400187 0.5518573
#> 7  0.62746580 0.6322365 0.6278868 0.63140419 0.6274396 0.6376519 0.6331390
#> 8  0.71716978 0.7264950 0.7209979 0.71730777 0.7229733 0.7229057 0.7195822
#> 9  0.80987226 0.8098160 0.8157056 0.80846493 0.8133580 0.8050364 0.8094571
#> 10 0.89933174 0.8934179 0.8966885 0.90414525 0.8995293 0.9016847 0.8927714
#>            X8         X9        X10       X11        X12       X13        X14
#> 1  0.09748547 0.09988567 0.09650701 0.1006151 0.09985985 0.1054672 0.09976289
#> 2  0.19460276 0.18573243 0.18805458 0.1964423 0.19205984 0.1899442 0.18469403
#> 3  0.28413348 0.28250174 0.27999151 0.2861555 0.28677286 0.2727816 0.27666681
#> 4  0.37171076 0.36472419 0.36425980 0.3633649 0.37228359 0.3773667 0.37113730
#> 5  0.45596371 0.45631656 0.45492162 0.4557941 0.45662222 0.4648776 0.45339441
#> 6  0.54609530 0.53844216 0.54225439 0.5445046 0.54398150 0.5410323 0.53997841
#> 7  0.63979527 0.63021669 0.63312519 0.6280282 0.63260796 0.6329995 0.64092734
#> 8  0.71963378 0.72742987 0.72438127 0.7261787 0.71903443 0.7163944 0.72217729
#> 9  0.80584744 0.81169714 0.80827189 0.8149406 0.80858341 0.8021204 0.80854583
#> 10 0.89913547 0.90560540 0.90699952 0.9005675 0.90175157 0.9020619 0.89760501
#>          X15
#> 1  0.1039365
#> 2  0.1872397
#> 3  0.2752245
#> 4  0.3700380
#> 5  0.4542282
#> 6  0.5400721
#> 7  0.6356751
#> 8  0.7246726
#> 9  0.8113128
#> 10 0.9006465