Plot the variables from the results of the cond_maha function.

Plot the variables from the results of the cond_maha function.

Usage

# S3 method for class 'cond_maha'
plot(
  x,
  ...,
  p_tail = 0,
  family = "sans",
  score_digits = ifelse(min(x$sigma) >= 10, 0, 2)
)

Arguments

x

The results of the cond_maha function.

...

Arguments passed to print function

p_tail

The proportion of the tail to shade

family

Font family.

score_digits

Number of digits to round scores.

Value

A ggplot2-object