Create a tibble to hold everything for simulation

Create a tibble to hold everything for simulation

simMaha(dPar, SimModel, n = 100)

Arguments

dPar
Parameters we would like to test in simulation studies

SimModel
Population relations among variables represented by Lavaan Syntax

n
The number of cases

Value

all parameters, lavaan object with specified parameters and data genereged for each condition by calling simStandarized

Examples

SimModel <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
makePara(1000, 9000, n = 1000, k = 15)->beta_measure1
makePara(1000, 1000, n = 1000, k = 8)->beta_structure1
cbind(beta_measure1, beta_structure1)-> par_beta1
colnames(par_beta1) <- paste0("x", 1:23)
simMaha(par_beta1, SimModel, n = 10)

#> # A tibble: 1,000 x 25
#>        x1    x2     x3    x4    x5     x6     x7     x8     x9    x10    x11
#>     <dbl> <dbl>  <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#>  1 0.103  0.104 0.0997 0.100 0.101 0.103  0.101  0.102  0.0959 0.0994 0.0972
#>  2 0.103  0.104 0.106  0.102 0.102 0.106  0.104  0.102  0.101  0.0990 0.101 
#>  3 0.101  0.103 0.104  0.102 0.103 0.0991 0.108  0.104  0.0978 0.102  0.102 
#>  4 0.0983 0.104 0.105  0.101 0.102 0.104  0.0988 0.0982 0.103  0.107  0.0960
#>  5 0.106  0.102 0.102  0.103 0.103 0.102  0.103  0.104  0.103  0.103  0.107 
#>  6 0.105  0.105 0.0986 0.104 0.108 0.105  0.103  0.103  0.104  0.0997 0.106 
#>  7 0.108  0.110 0.109  0.103 0.108 0.107  0.107  0.101  0.109  0.105  0.0995
#>  8 0.104  0.108 0.104  0.103 0.109 0.102  0.101  0.103  0.103  0.108  0.108 
#>  9 0.109  0.108 0.106  0.104 0.111 0.108  0.103  0.108  0.103  0.102  0.105 
#> 10 0.109  0.109 0.109  0.103 0.109 0.108  0.103  0.107  0.103  0.110  0.110 
#> # ... with 990 more rows, and 14 more variables: x12 <dbl>, x13 <dbl>,
#> #   x14 <dbl>, x15 <dbl>, x16 <dbl>, x17 <dbl>, x18 <dbl>, x19 <dbl>,
#> #   x20 <dbl>, x21 <dbl>, x22 <dbl>, x23 <dbl>, .out <list>, data <list>