Function that takes a lavaan model with standardized parameters and simulates latent scores, errors, disturbances, and observed scores

Function that takes a lavaan model with standardized parameters and simulates latent scores, errors, disturbances, and observed scores

simStandardized(m, n = 1000, ObservedOnly = FALSE)

Arguments

m
Structural model represented by lavaan Syntax

n
Number of simulated cases

ObservedOnly
Return only observed data

Value

Latent scores, errors, disturbances, and observed scores

Examples

# Lavaan model
m = "Latent_1 =~ 0.8 * Ob_1 + 0.8 * Ob_2"

# simulate 10 cases
simStandardized(m, n = 10)

#> $Data
#>      Latent_1          Ob_1        Ob_2     e_Ob_1      e_Ob_2 Latent_1_FS
#> 1  -1.6979356 -1.2322417187 -0.63035386  0.1261068  0.72799462  -0.9085832
#> 2   0.2827422  0.4851303533  0.63714377  0.2589366  0.41094999   0.5474508
#> 3   0.5372324  0.5591502340  0.23139205  0.1293643 -0.19839386   0.3856304
#> 4  -1.2872789 -0.5530686638 -1.57165654  0.4767545 -0.54183340  -1.0364513
#> 5   0.5396882  0.0762645730  1.14173419 -0.3554860  0.70998359   0.5941457
#> 6  -1.0362650 -0.1682714284 -0.63884507  0.6607406  0.19016693  -0.3937154
#> 7   0.7806475  0.5442048664 -0.21156361 -0.0803131 -0.83608157   0.1622640
#> 8   0.7894043  1.2363100750 -0.12542552  0.6047866 -0.75694897   0.5418949
#> 9  -0.2016445 -0.6562433760 -0.07055059 -0.4949277  0.09076504  -0.3545336
#> 10 -0.9831427 -0.0002976762 -0.69044202  0.7862165  0.09607212  -0.3369462
#>     e_Ob_1_FS  e_Ob_2_FS
#> 1  -0.8422919  0.1608545
#> 2   0.0786162  0.3319719
#> 3   0.4177432 -0.1285204
#> 4   0.4601540 -1.2374925
#> 5  -0.6650867  1.1106960
#> 6   0.2445014 -0.5397880
#> 7   0.6906561 -0.5689581
#> 8   1.3379903 -0.9315691
#> 9  -0.6210274  0.3551272
#> 10  0.4487655 -0.7014751
#> 
#> $vObserved
#> [1] "Ob_1" "Ob_2"
#> 
#> $vError
#> [1] "e_Ob_1" "e_Ob_2"
#> 
#> $R
#>          Latent_1 Ob_1 Ob_2 e_Ob_1 e_Ob_2
#> Latent_1      1.0 0.80 0.80    0.0    0.0
#> Ob_1          0.8 1.00 0.64    0.6    0.0
#> Ob_2          0.8 0.64 1.00    0.0    0.6
#> e_Ob_1        0.0 0.60 0.00    1.0    0.0
#> e_Ob_2        0.0 0.00 0.60    0.0    1.0
#> 
#> $A
#>          Latent_1 Ob_1 Ob_2
#> Latent_1      0.0    0    0
#> Ob_1          0.8    0    0
#> Ob_2          0.8    0    0
#> 
#> $S
#>          Latent_1 Ob_1 Ob_2
#> Latent_1        1 0.00 0.00
#> Ob_1            0 0.36 0.00
#> Ob_2            0 0.00 0.36
#> 
#> $iterations
#> [1] 2
#> 
#> $vLatent
#> [1] "Latent_1"
#> 
#> $vDisturbance
#> character(0)
#> 
#> $vFactorScores
#> [1] "Latent_1_FS" "e_Ob_1_FS"   "e_Ob_2_FS"  
#> 
#> $FactorScoreCoef
#>       Latent_1     e_Ob_1     e_Ob_2
#> Ob_1 0.4878049  1.0162602 -0.6504065
#> Ob_2 0.4878049 -0.6504065  1.0162602
#> 
#> $R_all
#>              Latent_1      Ob_1      Ob_2     e_Ob_1     e_Ob_2 Latent_1_FS
#> Latent_1    1.0000000 0.8000000 0.8000000  0.0000000  0.0000000   0.8834522
#> Ob_1        0.8000000 1.0000000 0.6400000  0.6000000  0.0000000   0.9055385
#> Ob_2        0.8000000 0.6400000 1.0000000  0.0000000  0.6000000   0.9055385
#> e_Ob_1      0.0000000 0.6000000 0.0000000  1.0000000  0.0000000   0.3312946
#> e_Ob_2      0.0000000 0.0000000 0.6000000  0.0000000  1.0000000   0.3312946
#> Latent_1_FS 0.8834522 0.9055385 0.9055385  0.3312946  0.3312946   1.0000000
#> e_Ob_1_FS   0.3748170 0.7683749 0.0000000  0.7808688 -0.4997560   0.4242641
#> e_Ob_2_FS   0.3748170 0.0000000 0.7683749 -0.4997560  0.7808688   0.4242641
#>              e_Ob_1_FS  e_Ob_2_FS
#> Latent_1     0.3748170  0.3748170
#> Ob_1         0.7683749  0.0000000
#> Ob_2         0.0000000  0.7683749
#> e_Ob_1       0.7808688 -0.4997560
#> e_Ob_2      -0.4997560  0.7808688
#> Latent_1_FS  0.4242641  0.4242641
#> e_Ob_1_FS    1.0000000 -0.6400000
#> e_Ob_2_FS   -0.6400000  1.0000000
#> 
#> $R_FS
#>           Latent_1     e_Ob_1     e_Ob_2
#> Latent_1 1.0000000  0.4242641  0.4242641
#> e_Ob_1   0.4242641  1.0000000 -0.6400000
#> e_Ob_2   0.4242641 -0.6400000  1.0000000
#> 
#> $FactorScoreValidity
#>  Latent_1    e_Ob_1    e_Ob_2 
#> 0.7804878 0.6097561 0.6097561 
#> 
#> $FactorScoreSE
#> se.Latent_1   se.e_Ob_1   se.e_Ob_2 
#>   0.4685213   0.6246950   0.6246950 
#> 
#> $Model
#> [1] "Latent_1 =~ 0.8 * Ob_1 + 0.8 * Ob_2"
#>