Wrapper for finding out Mahalanobis distance between variables: this one gives everything for practitioners to use when they only have population relations and their clients' data

Wrapper for finding out Mahalanobis distance between variables: this one gives everything for practitioners to use when they only have population relations and their clients' data

unusualness(data, model, v_dep, v_ind)

Arguments

data
Profiles of interest.

model
Structural model represented by lavaan Syntax.

v_dep
The names of variables you would like to condition on.

v_ind
The names of variables of your interest.

Value

conditional Mahalanobis distance, percentiles for each case based on the Chi-square distribution formed by conditional Mahalanobis distance and predicted Deps based on Inds.

Examples

# Standardized structural model in lavaan syntax
m <- "
Gc =~ 0.85 * Gc_1 + 0.68 * Gc_2 + 0.80 * Gc_3
Gf =~ 0.80 * Gf_1 + 0.90 * Gf_2 + 0.80 * Gf_3
Read =~ 0.66 * Read_1 + 0.85 * Read_2 + 0.91 * Read_3
Math =~ 0.40 * Math_1 + 0.90 * Math_2 + 0.70 * Math_3
Gc ~ 0.60 * Gf
Read ~ 0.40 * Gc + 0.10 * Gf
Math ~ 0.20 * Gc + 0.30 * Gf
"

# Put observed scores in data.frame
d_demo <- data.frame(
          Gc_1 = -1,
          Gc_2 = 0.5,
          Gc_3 = -0.2,
          Gf_1 = 1.1,
          Gf_2 = 1.3,
          Gf_3 = 2,
          Read_1 = -0.5,
          Read_2 = -1,
          Read_3 = -1.4,
          Math_1 = 1.1,
          Math_2 = 1.3,
          Math_3 = 0.7
)
unusualness(
     data = d_demo,
     model = m,
     v_dep = c("Math", "Read"),
     v_ind = c("Gc", "Gf"))

#> # A tibble: 5 x 7
#>   Measure      dCM `dCM Percentile` `dCM Level`    dM `dM Percentile` `dM Level`
#>   <chr>      <dbl>            <dbl> <chr>       <dbl>           <dbl> <chr>     
#> 1 "Latent D~  1.65           0.744  Average      1.87           0.827 High Aver~
#> 2 "Observed~  2.07           0.360  Average      2.24           0.457 Average   
#> 3 "Observed~  1.23           0.0421 Low          2.24           0.457 Average   
#> 4 "Observed~  2.28           0.482  Average      3.01           0.828 High Aver~
#> 5 "All Obse~  2.59           0.125  Low Average  3.65           0.653 Average