Calculate the conditional Mahalanobis distance for any variables.

Calculate the conditional Mahalanobis distance for any variables.

CMahalanobis(d, R, Dep, Ind = NULL, UseFactorScores_Ind = F,
  UseFactorScores_Dep = F, IncludeDiagnostics = F)

Arguments

d
Data.frame with the independent and dependent variables.

R
Conditional correlation among variables.

Dep
Vector of names of the dependent variables in your profile.

Ind
Vector of names of independent variables you would like to control for.

UseFactorScores_Ind
Use the factor scores for the independent variables

UseFactorScores_Dep
Use the factor scores for the dependent variables

IncludeDiagnostics
Return additional diagnostic information

Value

conditional Mahalanobis distance, percentiles for each case based on the Chi-square distribution formed by conditional Mahalanobis distance and predicted Deps based on Inds.

Examples

m <- "
Gc =~ 0.85 * Gc1 + 0.68 * Gc2 + 0.8 * Gc3
Gf =~ 0.8 * Gf1 + 0.9 * Gf2 + 0.8 * Gf3
Gs =~ 0.7 * Gs1 + 0.8 * Gs2 + 0.8 * Gs3
Read =~ 0.66 * Read1 + 0.85 * Read2 + 0.91 * Read3
Math =~ 0.4 * Math1 + 0.9 * Math2 + 0.7 * Math3
Gc ~ 0.6 * Gf + 0.1 * Gs
Gf ~ 0.5 * Gs
Read ~ 0.4 * Gc + 0.1 * Gf
Math ~ 0.2 * Gc + 0.3 * Gf + 0.1 * Gs"
d_demo <- simStandardized(m = m, n = 10)
CMahalanobis(d = d_demo$Data,
             R = d_demo$R_all,
             Dep = c("Math", "Read"),
             Ind = c("Gf", "Gs", "Gc"))

#> $dCM
#>  [1] 0.4848105 0.6641013 0.4505867 1.5435822 1.5867385 1.9697065 0.3043452
#>  [8] 0.7192421 0.5072176 1.2575261
#> 
#> $df
#> [1] 2
#> 
#> $p
#>  [1] 0.11087783 0.19789462 0.09653163 0.69618097 0.71602512 0.85627764
#>  [7] 0.04525692 0.22791036 0.12070500 0.54646644
#>