Estimate factot scores for a given profile and population model.

Estimate factot scores for a given profile and population model.

estStandardized(m, d)

Arguments

m
Lavaan Syntax Object.

d
Observed Z scores.

Value

Standardized estimated factor scores.

Examples

m = "LatantVariable =~ 0.8 * ObservedVar1 + 0.8 * ObservedVar2 + 0.8 * Observed3"
d <- simStandardized(m, 1000) #simulate 1000 cases (profiles)

# Pick up the first three cases for illustration
demo_individual_exp <- d$Data[1:3,]

# Estimate factor scores based on the three cases as well as the population model in m
estStandardized(m,  demo_individual_exp)

#> $Data
#>   LatantVariable ObservedVar1 ObservedVar2 Observed3 e_ObservedVar1
#> 1     1.57576437  1.495417788   2.28012710 2.0488545     0.23480629
#> 2     0.06493015 -0.003167523   0.07813167 0.5672660    -0.05511165
#> 3     0.42170339 -0.477339440   0.72223253 0.1772282    -0.81470216
#>   e_ObservedVar2 e_Observed3 LatantVariable_FS e_ObservedVar1_FS
#> 1     1.01951560   0.7882430         2.0436489        -0.2325022
#> 2     0.02618755   0.5153219         0.2253439        -0.3057377
#> 3     0.38486982  -0.1601346         0.1481127        -0.9930494
#>   e_ObservedVar2_FS e_Observed3_FS LatantVariable_FS e_ObservedVar1_FS
#> 1         1.0753466     0.68989227         2.0436489        -0.2325022
#> 2        -0.1702391     0.64498477         0.2253439        -0.3057377
#> 3         1.0062373     0.09789665         0.1481127        -0.9930494
#>   e_ObservedVar2_FS e_Observed3_FS
#> 1         1.0753466     0.68989227
#> 2        -0.1702391     0.64498477
#> 3         1.0062373     0.09789665
#> 
#> $vObserved
#> [1] "ObservedVar1" "ObservedVar2" "Observed3"   
#> 
#> $vError
#> [1] "e_ObservedVar1" "e_ObservedVar2" "e_Observed3"   
#> 
#> $vLatent
#> [1] "LatantVariable"
#> 
#> $vDisturbance
#> character(0)
#> 
#> $vFactorScores
#> [1] "LatantVariable_FS" "e_ObservedVar1_FS" "e_ObservedVar2_FS"
#> [4] "e_Observed3_FS"   
#> 
#> $FactorScoreCoef
#>              LatantVariable e_ObservedVar1 e_ObservedVar2 e_Observed3
#> ObservedVar1      0.3508772      1.1988304     -0.4678363  -0.4678363
#> ObservedVar2      0.3508772     -0.4678363      1.1988304  -0.4678363
#> Observed3         0.3508772     -0.4678363     -0.4678363   1.1988304
#> 
#> $R
#>                LatantVariable ObservedVar1 ObservedVar2 Observed3
#> LatantVariable            1.0         0.80         0.80      0.80
#> ObservedVar1              0.8         1.00         0.64      0.64
#> ObservedVar2              0.8         0.64         1.00      0.64
#> Observed3                 0.8         0.64         0.64      1.00
#> e_ObservedVar1            0.0         0.60         0.00      0.00
#> e_ObservedVar2            0.0         0.00         0.60      0.00
#> e_Observed3               0.0         0.00         0.00      0.60
#>                e_ObservedVar1 e_ObservedVar2 e_Observed3
#> LatantVariable            0.0            0.0         0.0
#> ObservedVar1              0.6            0.0         0.0
#> ObservedVar2              0.0            0.6         0.0
#> Observed3                 0.0            0.0         0.6
#> e_ObservedVar1            1.0            0.0         0.0
#> e_ObservedVar2            0.0            1.0         0.0
#> e_Observed3               0.0            0.0         1.0
#> 
#> $R_all
#>                   LatantVariable ObservedVar1 ObservedVar2 Observed3
#> LatantVariable               1.0         0.80         0.80      0.80
#> ObservedVar1                 0.8         1.00         0.64      0.64
#> ObservedVar2                 0.8         0.64         1.00      0.64
#> Observed3                    0.8         0.64         0.64      1.00
#> e_ObservedVar1               0.0         0.60         0.00      0.00
#> e_ObservedVar2               0.0         0.00         0.60      0.00
#> e_Observed3                  0.0         0.00         0.00      0.60
#> LatantVariable_FS            1.0         0.80         0.80      0.80
#> e_ObservedVar1_FS            0.0         0.60         0.00      0.00
#> e_ObservedVar2_FS            0.0         0.00         0.60      0.00
#> e_Observed3_FS               0.0         0.00         0.00      0.60
#>                   e_ObservedVar1 e_ObservedVar2 e_Observed3 LatantVariable_FS
#> LatantVariable               0.0            0.0         0.0               1.0
#> ObservedVar1                 0.6            0.0         0.0               0.8
#> ObservedVar2                 0.0            0.6         0.0               0.8
#> Observed3                    0.0            0.0         0.6               0.8
#> e_ObservedVar1               1.0            0.0         0.0               0.0
#> e_ObservedVar2               0.0            1.0         0.0               0.0
#> e_Observed3                  0.0            0.0         1.0               0.0
#> LatantVariable_FS            0.0            0.0         0.0               1.0
#> e_ObservedVar1_FS            1.0            0.0         0.0               0.0
#> e_ObservedVar2_FS            0.0            1.0         0.0               0.0
#> e_Observed3_FS               0.0            0.0         1.0               0.0
#>                   e_ObservedVar1_FS e_ObservedVar2_FS e_Observed3_FS
#> LatantVariable                  0.0               0.0            0.0
#> ObservedVar1                    0.6               0.0            0.0
#> ObservedVar2                    0.0               0.6            0.0
#> Observed3                       0.0               0.0            0.6
#> e_ObservedVar1                  1.0               0.0            0.0
#> e_ObservedVar2                  0.0               1.0            0.0
#> e_Observed3                     0.0               0.0            1.0
#> LatantVariable_FS               0.0               0.0            0.0
#> e_ObservedVar1_FS               1.0               0.0            0.0
#> e_ObservedVar2_FS               0.0               1.0            0.0
#> e_Observed3_FS                  0.0               0.0            1.0
#> 
#> $R_FS
#>                LatantVariable e_ObservedVar1 e_ObservedVar2 e_Observed3
#> LatantVariable      1.0000000      0.2705009      0.2705009   0.2705009
#> e_ObservedVar1      0.2705009      1.0000000     -0.3902439  -0.3902439
#> e_ObservedVar2      0.2705009     -0.3902439      1.0000000  -0.3902439
#> e_Observed3         0.2705009     -0.3902439     -0.3902439   1.0000000
#> 
#> $FactorScoreValidity
#> LatantVariable e_ObservedVar1 e_ObservedVar2    e_Observed3 
#>      0.8421053      0.7192982      0.7192982      0.7192982 
#> 
#> $FactorScoreSE
#> se.LatantVariable se.e_ObservedVar1 se.e_ObservedVar2    se.e_Observed3 
#>         0.3973597         0.5298129         0.5298129         0.5298129 
#>