Wrapper for finding out Mahalanobis distance between variables: this one gives everything for practitioners to use when they only have population relations and their clients' data

Wrapper for finding out Mahalanobis distance between variables: this one gives everything for practitioners to use when they only have population relations and their clients' data

maha(d, m, Dep, Ind = NULL)

Arguments

d
Profiles of interest.

m
Structural model represented by lavaan Syntax.

Dep
The names of variables you would like to condition on.

Ind
The names of variables of your interest.

Value

conditional Mahalanobis distance, percentiles for each case based on the Chi-square distribution formed by conditional Mahalanobis distance and predicted Deps based on Inds.

Examples

# Standardized structural model in lavaan syntax
m <- "
Gc =~ 0.85 * Gc_1 + 0.68 * Gc_2 + 0.80 * Gc_3
Gf =~ 0.80 * Gf_1 + 0.90 * Gf_2 + 0.80 * Gf_3
Read =~ 0.66 * Read_1 + 0.85 * Read_2 + 0.91 * Read_3
Math =~ 0.40 * Math_1 + 0.90 * Math_2 + 0.70 * Math_3
Gc ~ 0.60 * Gf
Read ~ 0.40 * Gc + 0.10 * Gf
Math ~ 0.20 * Gc + 0.30 * Gf
"

# Put observed scores in data.frame
d_demo <- data.frame(
          Gc_1 = -1,
          Gc_2 = 0.5,
          Gc_3 = -0.2,
          Gf_1 = 1.1,
          Gf_2 = 1.3,
          Gf_3 = 2,
          Read_1 = -0.5,
          Read_2 = -1,
          Read_3 = -1.4,
          Math_1 = 1.1,
          Math_2 = 1.3,
          Math_3 = 0.7
)
maha(d = d_demo,
     m = m,
     Dep = c("Math", "Read"),
     Ind = c("Gc", "Gf"))

#> $dCM
#> [1] 1.649632
#> 
#> $df
#> [1] 2
#> 
#> $p
#> [1] 0.7435041
#>